Initial commit
Uses 7DRL2017 as a base, because I didn't have the foresight to actually
clean anything up between then and now and so I'm just gonna have to deal
as best I can.

Killing the commit history by deleting the .git/ because it was, like, 350
commits long and full of throwaway, terrible, no-good very bad commit
messages since it was being done as a 7DRL and I was not worried about
maintainability or history. They'd just be very confusing.